# Decrease ClasspathStatementLocator GC pressure.