--- a/src/main/java/org/skife/jdbi/v2/ClasspathStatementLocator.java
+++ b/src/main/java/org/skife/jdbi/v2/ClasspathStatementLocator.java
@@ -22,6 +22,7 @@ import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 
 import java.io.IOException;
 import java.io.InputStream;
+import java.lang.reflect.Method;
 import java.util.Collections;
 import java.util.Map;
 import java.util.WeakHashMap;
@@ -33,7 +34,41 @@ import java.util.regex.Matcher;
  */
 public class ClasspathStatementLocator implements StatementLocator
 {
-    private final Map<Object, String> found = Collections.synchronizedMap(new WeakHashMap<Object, String>());
+    private static class CacheKey {
+        final String name;
+        final Class<?> sqlObjectType;
+        final Method sqlObjectMethod;
+
+        public CacheKey(String name, Class<?> sqlObjectType, Method sqlObjectMethod) {
+            this.name = name;
+            this.sqlObjectType = sqlObjectType;
+            this.sqlObjectMethod = sqlObjectMethod;
+        }
+
+        @Override
+        public boolean equals(Object o) {
+            if (this == o) return true;
+            if (o == null || getClass() != o.getClass()) return false;
+
+            CacheKey cacheKey = (CacheKey) o;
+
+            if (name != null ? !name.equals(cacheKey.name) : cacheKey.name != null) return false;
+            if (sqlObjectType != null ? !sqlObjectType.equals(cacheKey.sqlObjectType) : cacheKey.sqlObjectType != null)
+                return false;
+            return sqlObjectMethod != null ? sqlObjectMethod.equals(cacheKey.sqlObjectMethod) : cacheKey.sqlObjectMethod == null;
+
+        }
+
+        @Override
+        public int hashCode() {
+            int result = name != null ? name.hashCode() : 0;
+            result = 31 * result + (sqlObjectType != null ? sqlObjectType.hashCode() : 0);
+            result = 31 * result + (sqlObjectMethod != null ? sqlObjectMethod.hashCode() : 0);
+            return result;
+        }
+    }
+
+    private final Map<CacheKey, String> found = Collections.synchronizedMap(new WeakHashMap<CacheKey, String>());
 
     /**
      * Very basic sanity test to see if a string looks like it might be sql
@@ -75,19 +110,9 @@ public class ClasspathStatementLocator implements StatementLocator
     @SuppressFBWarnings("DM_STRING_CTOR")
     public String locate(String name, StatementContext ctx)
     {
-        final Object cache_key;
-        boolean isSqlObjectMethod = false;
-        if (ctx.getSqlObjectType() != null) {
-            if (ctx.getSqlObjectMethod() != null) {
-                isSqlObjectMethod = true;
-                cache_key = ctx.getSqlObjectMethod();
-            } else {
-                cache_key = '/' + mungify(ctx.getSqlObjectType().getName() + '.' + name) + ".sql";
-            }
-        }
-        else {
-            cache_key = name;
-        }
+        final CacheKey cache_key = new CacheKey(name, ctx.getSqlObjectType(), ctx.getSqlObjectMethod());
+        boolean isSqlObjectMethod = ctx.getSqlObjectType() != null && ctx.getSqlObjectMethod() != null;
+
 
         String cached = found.get(cache_key);
         if (cached != null) {
@@ -122,7 +147,7 @@ public class ClasspathStatementLocator implements StatementLocator
             if (in_stream == null) {
                 // Ensure we don't store an identity map entry which has a hard reference
                 // to the key (through the value) by copying the value, avoids potential memory leak.
-                found.put(cache_key, name == cache_key ? new String(name) : name);
+                found.put(cache_key, isSqlObjectMethod ? name : new String(name));
                 return name;
             }
             String sql;