Hi!

ClasspathStatementLocator gives essential GC pressure due to many string concatenation for every query even if static SQL objects are used.
```
cache_key = '/' + mungify(ctx.getSqlObjectType().getName() + '.' + name) + ".sql";
```
I propose to use Method object as cache key, this is completely enough for SQL objects.

And also I propose to cache individual SQL statements for SQL object methods cause they are static.