@qualidafial 2016-07-23
Hi @john9x! Thanks for your observations, and for submitting a PR.

Have you done any before-and-after measurements to validate that this change actually reduces GC pressure?