@john9x 2016-07-23
I can't give you concrete measurements right now.
I observed GC pressure from ClasspathStatmentLocator during profiling my application in the production environment.
In my case best solution was to replace ClasspathStatementLocator with stub. 

>Is there ever a case where sqlObjectType is set, but sqlObjectMethod is not?

It is just for any case to not get strange NPE.