@john9x 2016-07-24
I wrote a naive synthetic test but it shows the difference in GC pressure well.
JDK 1.8.0_60

VM options
`-Xmx16m -verbose:gc -Xloggc:improved.gc.log -XX:+PrintGCDetails -XX:+PrintGCDateStamps`
After ten minute run for old and new versions I loaded GC logs to [GCViewer](https://github.com/chewiebug/GCViewer).

Old version:
Total freed memory: 255M
Number of GC pauses: 84

Improved version:
Total freed memory: 35M
Numer of GC pauses: 11

```
public class ImprovedCslTest {
    public static void main(String[] args) throws Exception {
        int threadsCount = 100;
        List<Thread> threads = new ArrayList<Thread>(threadsCount);
        final StatementLocator csl = new ClasspathStatementLocator();
        final StatementContext ctx1 = new TestingStatementContext(new HashMap<String, Object>()) {
            @Override
            public Class<?> getSqlObjectType() {
                return ImprovedCslTest.class;
            }
            @Override
            public Method getSqlObjectMethod() {
                try {
                    return ImprovedCslTest.class.getMethod("testMethod1");
                } catch (NoSuchMethodException e) {
                    throw new RuntimeException(e);
                }
            }
        };
        final StatementContext ctx2 = new TestingStatementContext(new HashMap<String, Object>()) {
            @Override
            public Class<?> getSqlObjectType() {
                return ImprovedCslTest.class;
            }

            @Override
            public Method getSqlObjectMethod() {
                try {
                    return ImprovedCslTest.class.getMethod("testMethod2");
                } catch (NoSuchMethodException e) {
                    throw new RuntimeException(e);
                }
            }
        };
        final String sql1 = "SELECT some_column, other_column, one_more_column FROM table JOIN USING(id) WHERE a=1 AND b=2 OR c=3 ORDER BY d LIMIT 100500";
        final String sql2 = "INSERT INTO some_table some_column, other_column, one_more_column FROM table JOIN USING(id) WHERE a=1 AND b=2 OR c=3 ORDER BY d LIMIT 100500";
        final AtomicInteger cnt = new AtomicInteger();
        for (int i = 0; i < threadsCount; i++) {
            Thread t = new Thread(new Runnable() {
                @Override
                public void run() {
                    try {
                        Thread.sleep(new Random().nextInt(1000));
                        for (; ; ) {
                            Thread.sleep(1000);
                            if (new Random().nextDouble() < 0.5) {
                                csl.locate(sql1, ctx1);
                            } else {
                                csl.locate(sql2, ctx2);
                            }
                            cnt.incrementAndGet();
                        }

                    } catch (Exception e) {
                        e.printStackTrace();
                    }
                }
            });
            t.setDaemon(true);
            threads.add(t);
        }
        for (Thread thread : threads) {
            thread.start();
        }
        Thread.sleep(10 * 60 * 1000);
        System.gc();
        System.out.println("Calls: " + cnt.intValue());
    }
    public void testMethod1() {   }
    public void testMethod2() {   }
}
```