@qualidafial 2016-07-24
:+1: from me.

@stevenschlansker or @arteam, any concerns about this merge? I think using the `Method` as the cache key should be safe, it's the only part I was concerned about but I think the logic is sound.

If we're good to merge this, we should also forward these changes to the v3 branch, as it performs a similar "munge" operation in the new SQL Locators.