@arteam 2016-07-25
I've wrote a simple microbenchmark:
```java
@BenchmarkMode(Mode.Throughput)
@OutputTimeUnit(TimeUnit.MILLISECONDS)
public class QueriesBenchmark {

    @State(Scope.Benchmark)
    public static class DB {

        Dao dao;

        public DB() {
            Handle h = new DBI("jdbc:h2:mem:queries-benchmark", "sa", "").open();
            h.execute("create table something (id int primary key, name varchar(100))");

            PreparedBatch batch = h.prepareBatch("insert into something (id, name) values (?,?)");
            batch.add(1, "Brian");
            batch.add(2, "Thom");
            batch.execute();

            dao = h.attach(Dao.class);
        }
    }

    interface Dao {

        @SqlQuery("select name from something order by id")
        List<String> query();
    }

    @Benchmark
    public List<String> measureQuery(DB db) {
        return db.dao.query();
    }


    public static void main(String[] args) throws Exception {
        Options opt = new OptionsBuilder()
                .include(QueriesBenchmark.class.getSimpleName())
                .warmupIterations(10)
                .measurementIterations(5)
                .forks(1)
                .addProfiler(GCProfiler.class)
                .build();


        new Runner(opt).run();

    }
}
```

and got the following results:

Old version
```
Benchmark                                                        Mode  Cnt     Score      Error   Units
QueriesBenchmark.measureQuery                                   thrpt    5   234.525 ±    8.454  ops/ms
QueriesBenchmark.measureQuery:·gc.alloc.rate                    thrpt    5  1941.566 ±   68.475  MB/sec
QueriesBenchmark.measureQuery:·gc.alloc.rate.norm               thrpt    5  8683.776 ±   32.468    B/op
QueriesBenchmark.measureQuery:·gc.churn.PS_Eden_Space           thrpt    5  1952.571 ±  330.605  MB/sec
QueriesBenchmark.measureQuery:·gc.churn.PS_Eden_Space.norm      thrpt    5  8731.176 ± 1230.080    B/op
QueriesBenchmark.measureQuery:·gc.churn.PS_Survivor_Space       thrpt    5     0.112 ±    0.182  MB/sec
QueriesBenchmark.measureQuery:·gc.churn.PS_Survivor_Space.norm  thrpt    5     0.502 ±    0.812    B/op
QueriesBenchmark.measureQuery:·gc.count                         thrpt    5    53.000             counts
QueriesBenchmark.measureQuery:·gc.time                          thrpt    5    42.000                 ms
```

New version
```
Benchmark                                                        Mode  Cnt     Score      Error   Units
QueriesBenchmark.measureQuery                                   thrpt    5   240.153 ±  113.416  ops/ms
QueriesBenchmark.measureQuery:·gc.alloc.rate                    thrpt    5  1310.922 ±  621.074  MB/sec
QueriesBenchmark.measureQuery:·gc.alloc.rate.norm               thrpt    5  5728.005 ±    0.002    B/op
QueriesBenchmark.measureQuery:·gc.churn.PS_Eden_Space           thrpt    5  1306.050 ±  599.641  MB/sec
QueriesBenchmark.measureQuery:·gc.churn.PS_Eden_Space.norm      thrpt    5  5716.802 ± 1271.977    B/op
QueriesBenchmark.measureQuery:·gc.churn.PS_Survivor_Space       thrpt    5     0.069 ±    0.101  MB/sec
QueriesBenchmark.measureQuery:·gc.churn.PS_Survivor_Space.norm  thrpt    5     0.292 ±    0.321    B/op
QueriesBenchmark.measureQuery:·gc.count                         thrpt    5    37.000             counts
QueriesBenchmark.measureQuery:·gc.time                          thrpt    5    39.000                 ms
```

It looks like that the claim is correct, but it doesn't bring a big performance gain. 

Regarding the implementation, I am not really fond of mixing `String`s and `Method`s in one cache. That's not really obvious and could lead to strange errors. Probably we can split it to two caches?