@john9x 2016-07-25
Hmm. It looks like your benchmark for new version was not in a steady state. Too big mean error.
>QueriesBenchmark.measureQuery                                   thrpt    5   240.153 ±  113.416  ops/ms
QueriesBenchmark.measureQuery:·gc.alloc.rate                    thrpt    5  1310.922 ±  621.074  MB/sec
QueriesBenchmark.measureQuery:·gc.churn.PS_Eden_Space           thrpt    5  1306.050 ±  599.641  MB/sec
QueriesBenchmark.measureQuery:·gc.churn.PS_Eden_Space.norm      thrpt    5  5716.802 ± 1271.977    B/op

Your benchmark, my results:
```
# JMH 1.13 (released 2 days ago)
# VM version: JDK 1.8.0_60, VM 25.60-b23
# VM invoker: C:\Program Files\Java\jdk1.8.0_60\jre\bin\java.exe
# VM options: -Didea.launcher.port=7532 -Didea.launcher.bin.path=C:\Program Files (x86)\JetBrains\IntelliJ IDEA 15.0.3\bin -Dfile.encoding=UTF-8
# Warmup: 10 iterations, 1 s each
# Measurement: 5 iterations, 1 s each
# Timeout: 10 min per iteration
# Threads: 1 thread, will synchronize iterations
# Benchmark mode: Throughput, ops/time
# Benchmark: org.skife.jdbi.QueriesBenchmark.measureQuery

Old version:

Benchmark                                                        Mode  Cnt     Score      Error   Units
QueriesBenchmark.measureQuery                                   thrpt    5   188,665 ±    4,336  ops/ms
QueriesBenchmark.measureQuery:·gc.alloc.rate                    thrpt    5  1600,308 ±   36,740  MB/sec
QueriesBenchmark.measureQuery:·gc.alloc.rate.norm               thrpt    5  8896,006 ±    0,001    B/op
QueriesBenchmark.measureQuery:·gc.churn.PS_Eden_Space           thrpt    5  1606,154 ±  680,094  MB/sec
QueriesBenchmark.measureQuery:·gc.churn.PS_Eden_Space.norm      thrpt    5  8925,767 ± 3650,992    B/op
QueriesBenchmark.measureQuery:·gc.churn.PS_Survivor_Space       thrpt    5     0,056 ±    0,198  MB/sec
QueriesBenchmark.measureQuery:·gc.churn.PS_Survivor_Space.norm  thrpt    5     0,314 ±    1,106    B/op
QueriesBenchmark.measureQuery:·gc.count                         thrpt    5    59,000             counts
QueriesBenchmark.measureQuery:·gc.time                          thrpt    5    40,000                 ms

New version:

Benchmark                                                        Mode  Cnt     Score     Error   Units
QueriesBenchmark.measureQuery                                   thrpt    5   295,221 ±   1,312  ops/ms
QueriesBenchmark.measureQuery:·gc.alloc.rate                    thrpt    5  1598,867 ±   7,189  MB/sec
QueriesBenchmark.measureQuery:·gc.alloc.rate.norm               thrpt    5  5680,004 ±   0,001    B/op
QueriesBenchmark.measureQuery:·gc.churn.PS_Eden_Space           thrpt    5  1589,077 ± 145,234  MB/sec
QueriesBenchmark.measureQuery:·gc.churn.PS_Eden_Space.norm      thrpt    5  5645,252 ± 520,939    B/op
QueriesBenchmark.measureQuery:·gc.churn.PS_Survivor_Space       thrpt    5     0,188 ±   0,225  MB/sec
QueriesBenchmark.measureQuery:·gc.churn.PS_Survivor_Space.norm  thrpt    5     0,666 ±   0,800    B/op
QueriesBenchmark.measureQuery:·gc.count                         thrpt    5    89,000            counts
QueriesBenchmark.measureQuery:·gc.time                          thrpt    5    47,000                ms
```
This PR is not about huge performance gain but about GC pressure.
For new version - gc.alloc.rate.norm ~5680 B/op, for old version - 8896 B/op. 

>mixing Strings and Methods in one cache
IMHO it's not a big deal for private cache that isolated in one class.