@john9x 2016-07-25
And 1 second measurement time is not enough to get accurate results about GC.
10 seconds benchmark
```
Old version:

Benchmark                                                        Mode  Cnt     Score    Error   Units
QueriesBenchmark.measureQuery                                   thrpt    5   188,732 ±  0,533  ops/ms
QueriesBenchmark.measureQuery:·gc.alloc.rate                    thrpt    5  1595,513 ±  4,643  MB/sec
QueriesBenchmark.measureQuery:·gc.alloc.rate.norm               thrpt    5  8864,565 ±  1,015    B/op
QueriesBenchmark.measureQuery:·gc.churn.PS_Eden_Space           thrpt    5  1596,018 ±  7,266  MB/sec
QueriesBenchmark.measureQuery:·gc.churn.PS_Eden_Space.norm      thrpt    5  8867,368 ± 19,893    B/op
QueriesBenchmark.measureQuery:·gc.churn.PS_Survivor_Space       thrpt    5     0,120 ±  0,038  MB/sec
QueriesBenchmark.measureQuery:·gc.churn.PS_Survivor_Space.norm  thrpt    5     0,666 ±  0,208    B/op
QueriesBenchmark.measureQuery:·gc.count                         thrpt    5  2434,000           counts
QueriesBenchmark.measureQuery:·gc.time                          thrpt    5  1350,000               ms


New version:

Benchmark                                                        Mode  Cnt     Score    Error   Units
QueriesBenchmark.measureQuery                                   thrpt    5   298,840 ±  1,692  ops/ms
QueriesBenchmark.measureQuery:·gc.alloc.rate                    thrpt    5  1641,841 ±  9,184  MB/sec
QueriesBenchmark.measureQuery:·gc.alloc.rate.norm               thrpt    5  5761,021 ±  1,637    B/op
QueriesBenchmark.measureQuery:·gc.churn.PS_Eden_Space           thrpt    5  1643,517 ± 11,049  MB/sec
QueriesBenchmark.measureQuery:·gc.churn.PS_Eden_Space.norm      thrpt    5  5766,906 ± 43,958    B/op
QueriesBenchmark.measureQuery:·gc.churn.PS_Survivor_Space       thrpt    5     0,110 ±  0,037  MB/sec
QueriesBenchmark.measureQuery:·gc.churn.PS_Survivor_Space.norm  thrpt    5     0,386 ±  0,129    B/op
QueriesBenchmark.measureQuery:·gc.count                         thrpt    5  2267,000           counts
QueriesBenchmark.measureQuery:·gc.time                          thrpt    5  1310,000               ms
```